Add: 무기에 따른 몽타주, 공격 타입별 데미지, 파티클 이펙트  추가
무기 타입에 따른 공격 방식 및 몽타주를 추가

공격 타입에 따른 데미지를 구조체로 만들어 데미지 처리

차크라 시전, 차징 시를 알리는 시각적 요소인 파티클을 추가

+++ Source/Naruto/Private/HealthManager.cpp
@@ -15,7 +15,7 @@ void UHealthManager::TickComponent(float DeltaTime, ELevelTick TickType, FActorC
 	Super::TickComponent(DeltaTime, TickType, ThisTickFunction);
 
 }
-void UHealthManager::SetDecreaseHealth(float val) {
+void UHealthManager::SetDecreaseHealth(int8 val) {
 	CurrentHealth -= val;
 	SetHealthRatio();
 	//UE_LOG(LogTemp, Warning, TEXT("Hited %f damage -> %f"), val, CurrentHealth);
